Es wurden Features, wie Edit und Delete von Dienstleistungen für Companyuser hinzugefügt. Zusätzlich wurde der Login korrigiert, weil dieser den UserType nicht erkannte. Appointments Komponente wurde hinzugefügt und das Eingabefenster under Services für Appointments wurde eingerichtet. Jedoch funktioniert die Speicherung in der Datenbank noch nicht. Zusätzlich wurde die Profilseite errichtet, welches jedoch noch nicht fertig ist.

+++ src/main/java/bookNow/Requests/ServiceCompanyUpdate.java
@@ -3,6 +3,8 @@ package bookNow.Requests;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ServiceCompanyUpdate {
 
@@ -10,7 +12,7 @@ public class ServiceCompanyUpdate {
 
     private String description;
 
-    private int price;
+    private BigDecimal price;
 
     private int duration;
